Update jk_ff.v
From looking at the RTL code I believe that you used vivado. I found that you used clk in your if statement's argument. And I strongly believe that it is not a good practice for a synthesizable code. And also using case statement instead of if else is appropriate for this design as it doesnt need hierarchical mux but a flat mux. Look into the code in my pull request and let me know what you think. I checked the elaborated design for the new code and it work perfect. Reach out to me if you like to have a discussion. Good luck!